chore: archivos redistribuidos
