+++ .gitignore
@@ -1,3 +1,4 @@
 /node_modules
 /.idea
 *.tgz
+.vscode
