+++ lib/cursos/Curso.js
@@ -16,7 +16,7 @@ const Modulo = require('./Modulo')
 
 class Curso {
   // TODO: Forzar tipos de los parámetros.
-  constructor(nrc, sigla, permiteRetiro, ingles, seccion, aprobacionEspecial, area, formato, categoria, nombre,
+  constructor (nrc, sigla, permiteRetiro, ingles, seccion, aprobacionEspecial, area, formato, categoria, nombre,
     profesor, campus, creditos, vacantesDisponibles, horario) {
     this.nrc = nrc
     this.sigla = sigla
@@ -41,7 +41,7 @@ class Curso {
      * @param curso2
      * @returns {boolean}
      */
-  static mismoHorario(curso1, curso2) {
+  static mismoHorario (curso1, curso2) {
     const horario1 = curso1.horario
     const horario2 = curso2.horario
 
@@ -61,7 +61,7 @@ class Curso {
      * @param choquesPermitidos
      * @returns {boolean}
      */
-  static horariosCompatibles(curso1, curso2, choquesPermitidos) {
+  static horariosCompatibles (curso1, curso2, choquesPermitidos) {
     const horario1 = curso1.horario
     const horario2 = curso2.horario
     const sigla1 = curso1.sigla
@@ -87,7 +87,7 @@ class Curso {
      * @param horario
      * @returns {Curso}
      */
-  static cursoMinimo(nrc, sigla, seccion, nombre, profesor, vacantesDisponibles, horario) {
+  static cursoMinimo (nrc, sigla, seccion, nombre, profesor, vacantesDisponibles, horario) {
     return new Curso(nrc, sigla, undefined, undefined, seccion, undefined,
       '', '', '', nombre, profesor, '', 0, vacantesDisponibles, horario)
   }
