+++ lib/cursos/Modulo.js
@@ -14,7 +14,7 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
 class Modulo {
   // TODO: getter sin sentido acá
-  static get TIPOS() {
+  static get TIPOS () {
     return [
       'CLAS',
       'LAB',
@@ -27,7 +27,7 @@ class Modulo {
       'TES']
   }
 
-  static get DIAS() {
+  static get DIAS () {
     return [
       'L',
       'M',
@@ -37,12 +37,12 @@ class Modulo {
       'S']
   }
 
-  static get MODULOS() {
+  static get MODULOS () {
     return 8
   }
 
   // TODO: Forzar tipo de los parámetros.
-  constructor(tipo, dia, modulo, sala) {
+  constructor (tipo, dia, modulo, sala) {
     this.tipo = tipo
     this.dia = dia
     this.modulo = modulo
@@ -50,11 +50,11 @@ class Modulo {
     this.hora = modulo
   }
 
-  static modulosIdenticos(a, b) {
+  static modulosIdenticos (a, b) {
     return a.tipo === b.tipo && a.dia === b.dia && a.modulo === b.modulo
   }
 
-  static modulosCompatibles(a, b) {
+  static modulosCompatibles (a, b) {
     return (a.dia !== b.dia) || (a.modulo !== b.modulo) || (a.dia === 'SIN HORARIO') || (b.dia === 'SIN HORARIO')
   }
 }
